requirements.txt
blinker==1.8.2
click==8.1.7
docutils==0.21.2
cryptography==43.0.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lockfile==0.12.2
MarkupSafe==2.1.5
packaging==24.1
PyMySQL==1.1.1
pynput==1.7.7
# pyobjc-core==10.3.1
# pyobjc-framework-ApplicationServices==10.3.1
# pyobjc-framework-Cocoa==10.3.1
# pyobjc-framework-CoreText==10.3.1
# pyobjc-framework-Quartz==10.3.1
python-daemon==3.0.1
python-dotenv==1.0.1
setuptools==73.0.1
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.4
